requirements.txt
pendulum>=3.0.0,<4.0
numpy>=1.21.0
pandas>=1.3.0
selenium==4.10.0
webdriver-manager==4.0.2
pyspark==3.2.1
psycopg2-binary
requests
cython==0.29.24
streamlit==1.25.0
pyyaml
pyarrow
plotly
apache-airflow-providers-slack==5.1.0
apache-airflow-providers-openlineage>=2.0.0
apache-airflow-providers-ssh
seaborn
matplotlib
dotenv